+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Fixed issue with transaction toast title](https://github.com/multiversx/mx-sdk-dapp-ui/pull/200)
+
 ## [[0.0.25](https://github.com/multiversx/mx-sdk-dapp-ui/pull/199)] - 2025-08-12
 
 - [Fixed issue with transaction toast progress](https://github.com/multiversx/mx-sdk-dapp-ui/pull/198)