+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Removed Trim unrelated styling](https://github.com/multiversx/mx-sdk-dapp-ui/pull/201)
 - [Fixed issue with transaction toast title](https://github.com/multiversx/mx-sdk-dapp-ui/pull/200)
 
 ## [[0.0.25](https://github.com/multiversx/mx-sdk-dapp-ui/pull/199)] - 2025-08-12